Fix import loop
