+++ tasks/kernel_matrix_testing/platforms.py
@@ -1,0 +1,16 @@
+import json
+from typing import Dict, TypedDict, cast
+
+
+class Platforms(TypedDict):
+    url_base: str
+    x86_64: Dict[str, str]  # noqa: F841
+    arm64: Dict[str, str]  # noqa: F841
+
+
+platforms_file = "test/new-e2e/system-probe/config/platforms.json"
+
+
+def get_platforms():
+    with open(platforms_file) as f:
+        return cast(Platforms, json.load(f))
