+++ tasks/kernel_matrix_testing/vars.py
@@ -14,3 +14,6 @@ arch_mapping: Dict[str, Arch] = {
     "arm": "arm64",
     "aarch64": "arm64",
 }
+
+
+VMCONFIG = "vmconfig.json"
